feat: lint-cleanup-website-001 - Preserve dashboard Storybook states without file-size suppression
